+++ Src/AppLib/VpnHood.AppLib.Android.Common/Activities/AndroidAppWebViewMainActivityHandler.cs
@@ -36,10 +36,7 @@ public class AndroidAppWebViewMainActivityHandler(
     {
         try {
             if (!VpnHoodAppWebServer.IsInit) {
-                ArgumentNullException.ThrowIfNull(VpnHoodApp.Instance.Resources.SpaZipData);
-                using var spaZipStream = new MemoryStream(VpnHoodApp.Instance.Resources.SpaZipData);
                 VpnHoodAppWebServer.Init(new WebServerOptions {
-                    SpaZipStream = spaZipStream,
                     DefaultPort = options.SpaDefaultPort,
                     ListenOnAllIps = options.SpaListenToAllIps
                 });