+++ Src/AppLib/VpnHood.AppLib.WebServer/WebServerOptions.cs
@@ -2,7 +2,6 @@
 
 public class WebServerOptions
 {
-    public required Stream SpaZipStream { get; init; }
     public int? DefaultPort { get; init; }
     public Uri? Url { get; init; }
     public bool ListenOnAllIps { get; init; }