+++ Src/Apps/Client.Android.Google/App.cs
@@ -15,12 +15,11 @@ namespace VpnHood.App.Client.Droid.Google;
     SupportsRtl = AndroidAppConstants.SupportsRtl,
     AllowBackup = AndroidAppConstants.AllowBackup)]
 public class App(IntPtr javaReference, JniHandleOwnership transfer)
-    : VpnHoodAndroidApp(javaReference, transfer)
+    : Application(javaReference, transfer)
 {
-    protected override AppOptions CreateAppOptions()
+    private AppOptions CreateAppOptions()
     {
         var appConfigs = AppConfigs.Load();
-
         var resources = ClientAppResources.Resources;
         resources.Strings.AppName = AppConfigs.AppName;
 
@@ -39,7 +38,18 @@ public class App(IntPtr javaReference, JniHandleOwnership transfer)
                 UpdaterProvider = new GooglePlayAppUpdaterProvider(),
                 PromptDelay = TimeSpan.FromDays(3)
             },
-
         };
+    }
+
+    public override void OnCreate()
+    {
+        VpnHoodAndroidApp.Init(CreateAppOptions);
+        base.OnCreate();
+    }
+    
+    public override void OnTerminate()
+    {
+        if (VpnHoodAndroidApp.IsInit)
+            VpnHoodAndroidApp.Instance.Dispose();
     }
 }