+++ Src/Apps/Client.Linux.Web/Program.cs
@@ -1,74 +1,73 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using System.Diagnostics;
+using Microsoft.Extensions.Logging;
 using VpnHood.AppLib;
 using VpnHood.AppLib.Services.Updaters;
 using VpnHood.AppLib.WebServer;
-using VpnHood.Core.Client.Device.Linux;
+using VpnHood.Core.Proxies;
 using VpnHood.Core.Toolkit.Logging;
-using VpnHood.Core.Toolkit.Utils;
 
 namespace VpnHood.App.Client.Linux.Web;
 
 internal class Program
 {
-    private static async Task Main(string[] args)
+    private static AppOptions CreateAppOptions()
     {
-        _ = args;
-        try
-        {
-            Console.WriteLine("Starting VpnHood Client for linux (Beta).");
-            Console.WriteLine("On UI supported at this time.");
-            
-            var appConfigs = AppConfigs.Load();
+        var appConfigs = AppConfigs.Load();
+        var resources = ClientAppResources.Resources;
+        resources.Strings.AppName = AppConfigs.AppName;
+        var appOptions = new AppOptions(appConfigs.AppId, AppConfigs.AppName, AppConfigs.IsDebugMode) {
+            Resources = resources,
+            AccessKeys = AppConfigs.IsDebug ? [appConfigs.DefaultAccessKey] : [],
+            IsAddAccessKeySupported = true,
+            IsLocalNetworkSupported = true,
+            AllowEndPointStrategy = true,
+            DisconnectOnDispose = true,
+            AllowRecommendUserReviewByServer = false,
+            UpdaterOptions = new AppUpdaterOptions {
+                UpdateInfoUrl = appConfigs.UpdateInfoUrl,
+                PromptDelay = TimeSpan.Zero,
+            },
+            LogServiceOptions = {
+                SingleLineConsole = false
+            }
+        };
 
-            // try to remove the old adapter, as previous route may till be active
-            await VhUtils.TryInvokeAsync("Remove the old adapter", () =>
-                ExecuteCommandAsync($"ip link delete {AppConfigs.AppName}", CancellationToken.None));
-
-            // ReSharper disable once UseObjectOrCollectionInitializer
-            var resources = ClientAppResources.Resources;
-            resources.Strings.AppName = AppConfigs.AppName;
-
-            // check command line
-            // initialize VpnHoodApp
-            var appOptions = new AppOptions(appConfigs.AppId, AppConfigs.AppName, AppConfigs.IsDebugMode)
-            {
-                Resources = resources,
-                AccessKeys = AppConfigs.IsDebug ? [appConfigs.DefaultAccessKey] : [],
-                IsAddAccessKeySupported = true,
-                IsLocalNetworkSupported = true,
-                LocalSpaHostName = "my-vpnhood",
-                AllowEndPointStrategy = true,
-                AllowRecommendUserReviewByServer = false,
-                UpdaterOptions = new AppUpdaterOptions {
-                    UpdateInfoUrl = appConfigs.UpdateInfoUrl,
-                    PromptDelay = TimeSpan.Zero,
-                },
-                LogServiceOptions = {
-                    SingleLineConsole = false
-                }
-            };
-            VpnHoodApp.Init(new LinuxDevice(appOptions.StorageFolderPath), appOptions);
-
-            // initialize SPA
-            ArgumentNullException.ThrowIfNull(VpnHoodApp.Instance.Resources.SpaZipData);
-            using var spaZipStream = new MemoryStream(VpnHoodApp.Instance.Resources.SpaZipData);
-            var webServer = VpnHoodAppWebServer.Init(new WebServerOptions
-            {
-                SpaZipStream = spaZipStream
-            });
-
-            Console.WriteLine($"Navigate {webServer.Url} in your browser to open VpnHood Client.");
-            await Task.Delay(TimeSpan.FromMinutes(60));
-        }
-        catch (Exception ex)
-        {
-            VhLogger.Instance.LogError(ex, "Could not run the app.");
-            throw;
-        }
+        return appOptions;
     }
 
-    private static Task<string> ExecuteCommandAsync(string command, CancellationToken cancellationToken)
+    private static Task Main(string[] args)
     {
-        return OsUtils.ExecuteCommandAsync("/bin/bash", $"-c \"{command}\"", cancellationToken);
+        Console.WriteLine("Starting VpnHood Client for linux (Beta).");
+        Console.WriteLine("Only WebUI supported at this time.");
+
+        // init VpnHood app
+        VpnHoodAppLinux.Init(CreateAppOptions, args);
+
+        // init webserver
+        VpnHoodAppWebServer.Init(new WebServerOptions());
+        VpnHoodAppLinux.Instance.OpenMainWindowRequested += (_, _) => OpenMainWindow();
+
+        // run app
+        return VpnHoodAppLinux.Instance.Run();
+    }
+
+    private static void OpenMainWindow()
+    {
+        if (!VpnHoodAppWebServer.IsInit) {
+            VhLogger.Instance.LogWarning("VpnHood Web server is not initialized to show its web UI.");
+            return;
+        }
+
+        VhLogger.Instance.LogInformation("Running the default browser...");
+        Process.Start(new ProcessStartInfo {
+            FileName = "xdg-open",
+            ArgumentList = { VpnHoodAppWebServer.Instance.Url.AbsoluteUri },
+            UseShellExecute = false,
+            RedirectStandardError = true,
+            RedirectStandardOutput = true
+        });
+
+        VhLogger.Instance.LogInformation("To open VpnHood UI navigate to {0}", 
+            VpnHoodAppWebServer.Instance.Url);
     }
 }